atualizando a melhoria do dashboard 2 Qualidade

+++ src/models/Analise_2_qualidade.py
@@ -44,7 +44,8 @@ class Analise_2_qualidade():
         TotalPecas = tags['qtd'].sum()
 
         ordemProd = OrdemProd.OrdemProd(self.codEmpresa,'',self.data_inicio, self.data_final).ops_baixas_csw()
-        TotalPCsBaixadas = ordemProd['qtdMovto'].sum()
+
+        TotalPCsBaixadas = ordemProd['qtdMovto'].astype(int).sum()
 
         data = {
             '1- Peças com Motivo de 2Qual.': TotalPecas,
